pyproject.toml
[project]
name = "image-toolkit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.2",
    "opencv-python-headless>=4.13.0.92",
    "piexif>=1.1.3",
    "pillow>=12.1.1",
    "streamlit>=1.54.0",
]
